pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Olga <jasminess847@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"


[tool.poetry.group.dev.dependencies]
jinja2 = "^3.1.4"
flask = "^3.0.3"
flake8 = "^7.1.1"
python-dotenv = "^1.0.1"
gunicorn = "^23.0.0"
pytest-cov = "^5.0.0"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
